requirements.txt
beautifulsoup4==4.12.3
keyboard==0.13.5
paddleocr==2.8.1
paddlepaddle==3.0.0b0
Pillow==10.4.0
selenium==4.24.0
watchdog==5.0.0
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy<2.0
rapidfuzz
opencv-python
opencv-contrib-python
cython
Pillow
pyyaml
requests
albumentations==1.4.10
setuptools